New file: classes_functions.py
